requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
astroid==2.11.2
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
awscli==1.22.92
backports.weakref==1.0.post1
beautifulsoup4==4.11.1
black==22.3.0
botocore==1.24.37
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.3
cryptography==36.0.2
debugpy==1.6.0
dill==0.3.4
discord.py==2.0.1
docopt==0.6.2
docutils==0.15.2
frozenlist==1.3.1
google-auth==2.6.5
google-auth-oauthlib==0.5.1
greenlet==1.1.2
gspread==5.3.2
h11==0.13.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
isort==5.10.1
Jinja2==3.1.2
jmespath==1.0.0
lazy-object-proxy==1.7.1
lxml==4.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
MouseInfo==0.1.3
msgpack==1.0.3
multidict==6.0.2
mypy-extensions==0.4.3
neovim==0.3.1
oauth2client==4.1.3
oauthlib==3.2.0
outcome==1.1.0
pathspec==0.9.0
Pillow==9.1.0
pipreqs==0.4.11
platformdirs==2.5.1
progress==1.6
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pygame==2.1.2
PyGetWindow==0.0.9
pylint==2.13.5
PyMsgBox==1.0.9
pynvim==0.4.3
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-xlib==0.31
pytweening==1.0.4
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.7.2
rubicon-objc==0.4.2
s3transfer==0.5.2
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
toml==0.10.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.5.4
webencodings==0.5.1
wrapt==1.14.0
wsproto==1.1.0
yarg==0.1.9
yarl==1.8.1